refactor: Move cli config loading into Store
